get a bit

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -10,8 +10,8 @@ int get_bit(unsigned long int n, unsigned int index)
 {
 	int num = 1;
 
-	if (index >= (sizeof(unsigned int) * 8))
-			return (-1);
+	if (index >= (sizeof(unsigned long int) * 8))
+		return (-1);
 
 	if (n & (num << index))
 		return (num);
